extract the ConnectionPool interface
