+++ celesta-test/src/test/java/ru/curs/celesta/dbutils/adaptors/PostgresAdaptorTest.java
@@ -36,7 +36,7 @@ public class PostgresAdaptorTest extends AbstractAdaptorTest {
         cpc.setDriverClassName(appSettings.getDbClassName());
         cpc.setLogin(appSettings.getDBLogin());
         cpc.setPassword(appSettings.getDBPassword());
-        ConnectionPool connectionPool = ConnectionPool.create(cpc);
+        ConnectionPool connectionPool = InternalConnectionPool.create(cpc);
 
         dba = new PostgresAdaptor(connectionPool, new JdbcDdlConsumer());
